audioFile/requirements.txt
azure-functions
pydub
openai
